+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/StreamBindingService.java
@@ -16,9 +16,9 @@
 package com.expediagroup.streamplatform.streamregistry.core.services;
 
 import com.expediagroup.streamplatform.streamregistry.core.handlers.HandlerService;
-import com.expediagroup.streamplatform.streamregistry.core.view.ConsumerBindingView;
-import com.expediagroup.streamplatform.streamregistry.core.view.ProducerBindingView;
-import com.expediagroup.streamplatform.streamregistry.core.view.StreamBindingView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ConsumerBindingView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ProducerBindingView;
+import com.expediagroup.streamplatform.streamregistry.core.views.StreamBindingView;
 import com.expediagroup.streamplatform.streamregistry.core.validators.StreamBindingValidator;
 import com.expediagroup.streamplatform.streamregistry.core.validators.ValidationException;
 import com.expediagroup.streamplatform.streamregistry.model.Status;
