+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/ZoneService.java
@@ -21,7 +21,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Predicate;
 
-import com.expediagroup.streamplatform.streamregistry.core.view.ZoneView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ZoneView;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
