+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/validators/InfrastructureValidator.java
@@ -15,7 +15,7 @@
  */
 package com.expediagroup.streamplatform.streamregistry.core.validators;
 
-import com.expediagroup.streamplatform.streamregistry.core.view.ZoneView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ZoneView;
 import com.expediagroup.streamplatform.streamregistry.model.Infrastructure;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
