+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/validators/StreamBindingValidator.java
@@ -15,8 +15,8 @@
  */
 package com.expediagroup.streamplatform.streamregistry.core.validators;
 
-import com.expediagroup.streamplatform.streamregistry.core.view.InfrastructureView;
-import com.expediagroup.streamplatform.streamregistry.core.view.StreamView;
+import com.expediagroup.streamplatform.streamregistry.core.views.InfrastructureView;
+import com.expediagroup.streamplatform.streamregistry.core.views.StreamView;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Component;
