+++ core/src/test/java/com/expediagroup/streamplatform/streamregistry/core/services/ConsumerBindingServiceTest.java
@@ -22,7 +22,7 @@ import static org.mockito.Mockito.when;
 
 import java.util.Optional;
 
-import com.expediagroup.streamplatform.streamregistry.core.view.ConsumerBindingView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ConsumerBindingView;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
