+++ core/src/test/java/com/expediagroup/streamplatform/streamregistry/core/services/ConsumerServiceTest.java
@@ -22,8 +22,8 @@ import static org.mockito.Mockito.when;
 
 import java.util.Optional;
 
-import com.expediagroup.streamplatform.streamregistry.core.view.ConsumerBindingView;
-import com.expediagroup.streamplatform.streamregistry.core.view.ConsumerView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ConsumerBindingView;
+import com.expediagroup.streamplatform.streamregistry.core.views.ConsumerView;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
