+++ core/src/test/java/com/expediagroup/streamplatform/streamregistry/core/services/DomainServiceTest.java
@@ -23,7 +23,7 @@ import static org.mockito.Mockito.when;
 
 import java.util.Optional;
 
-import com.expediagroup.streamplatform.streamregistry.core.view.DomainView;
+import com.expediagroup.streamplatform.streamregistry.core.views.DomainView;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
